+++ packages/gi-assets-basic/src/deploys/index.ts
@@ -1,0 +1,5 @@
+import Esm from './Esm';
+import Html from './Html';
+import Umd from './Umd';
+
+export { Esm, Html, Umd };