+++ packages/gi-assets-scene/src/Timebar/types.ts
@@ -8,4 +8,6 @@ export type Selection = [string, string] | [number, number];
 
 export type TimeGranularity = 'year' | 'quarter' | 'month' | 'week' | 'day' | 'hour' | 'minute' | 'second' | number;
 
-export type TimeFiled = `${'nodes' | 'edges'}:${string}`;
+export type DataType = Record<string, any>[];
+
+export type FieldType = 'nodes' | 'edges';