+++ packages/gi-common-components/src/GroupContainer/ExpressionGroup.tsx
@@ -1,6 +1,12 @@
-import { DeleteOutlined, FieldNumberOutlined, FieldStringOutlined, PlusOutlined } from '@ant-design/icons';
-import { Button, Form, Col, Input, InputNumber, Row, Select } from 'antd';
-import React from 'react';
+import {
+  DeleteOutlined,
+  FieldNumberOutlined,
+  FieldStringOutlined,
+  FileTextOutlined,
+  PlusOutlined,
+} from '@ant-design/icons';
+import { Button, Col, Form, Input, InputNumber, Row, Select } from 'antd';
+import React, { useEffect, useState } from 'react';
 import { getOperatorList } from './utils';
 
 import './expressionGroup.less';
@@ -19,8 +25,33 @@ const ExpressionGroup: React.FunctionComponent<{
   name: string;
   options: any[];
 }> = ({ index, form, name, options }) => {
+  const [internalOptions, setInternalOptions] = useState(options);
+  // 支持自定义属性
+  const handleNameSearch = (value: string) => {
+    if (!value) {
+      setInternalOptions(options);
+      return;
+    }
+    const newOptions = options.filter(option => option.value.includes(value));
+    setInternalOptions([
+      {
+        label: value,
+        value,
+      },
+      ...newOptions,
+    ]);
+  };
+
+  const getPropertyType = (propertyName: string) => {
+    return options.find(option => option.value === propertyName)?.type || 'unknown';
+  };
+
+  useEffect(() => {
+    if (options?.length) setInternalOptions(options);
+  }, [options]);
+
   const content = (
-    <Form.Item name={[name, 'expressions']} label="属性过滤">
+    <Form.Item name={[name, 'expressions']} label="属性过滤" tooltip={'支持自定义属性路径，例: a.b.c'}>
       <Form.List name={[name, 'expressions']}>
         {(fields, { add, remove }) => {
           const { expressions } = form.getFieldValue('groups')[index] || {};
@@ -29,16 +60,23 @@ const ExpressionGroup: React.FunctionComponent<{
               {fields.map(({ key, name, ...restField }, conditionIndex) => {
                 const { name: property } = expressions[conditionIndex] || {};
                 // 根据类型判断 property 具体类型
-                const propertyType = (property || '').split('-')[1] || 'string';
+                const propertyType = getPropertyType(property);
                 return (
                   <Row align="middle" gutter={8} className="item" key={key}>
                     <Col span={2}>
                       {propertyType === 'string' && <FieldStringOutlined />}
-                      {(propertyType === 'long' || propertyType === 'double') && <FieldNumberOutlined />}
+                      {propertyType === 'number' && <FieldNumberOutlined />}
+                      {propertyType === 'unknown' && <FileTextOutlined />}
                     </Col>
                     <Col span={7}>
                       <Form.Item {...restField} name={[name, 'name']}>
-                        <Select size="small" style={{ width: '100%' }} options={options} />
+                        <Select
+                          size="small"
+                          showSearch
+                          onSearch={handleNameSearch}
+                          style={{ width: '100%' }}
+                          options={internalOptions}
+                        />
                       </Form.Item>
                     </Col>
                     <Col span={6}>