+++ packages/gi-sdk/src/GISDK.tsx
@@ -14,7 +14,7 @@ import type { Props, State } from './typing';
 import { GIComponentConfig } from './typing';
 import { createUuid } from './process/common';
 
-let updateHistoryTimer: NodeJS.Timer;
+let updateHistoryTimer: number;
 
 /** export  */
 const GISDK = (props: Props) => {
@@ -292,12 +292,15 @@ const GISDK = (props: Props) => {
     }
   };
 
+  const HAS_GRAPH = graphinRef.current?.graph && !graphinRef.current.graph.destroyed;
+
   const ContextValue = {
     ...state,
     GISDK_ID,
     services,
     assets,
     sourceDataMap,
+    HAS_GRAPH,
     graph: graphinRef.current?.graph,
     theme: graphinRef.current?.theme,
     apis: graphinRef.current?.apis,
@@ -344,8 +347,8 @@ const GISDK = (props: Props) => {
       };
       // 防止频繁更新导致的重复 updateHistory
       // 同时，间隔一定时间再更新到历史栈中，保证画布数据已经更新完成
-      if (updateHistoryTimer) clearTimeout(updateHistoryTimer);
-      updateHistoryTimer = setTimeout(fn, 500);
+      if (updateHistoryTimer) window.clearTimeout(updateHistoryTimer);
+      updateHistoryTimer = window.setTimeout(fn, 500);
     },
     stopForceSimulation: stopForceSimulation,
     restartForceSimulation: restartForceSimulation,
@@ -353,8 +356,6 @@ const GISDK = (props: Props) => {
   if (!ComponentAssets) {
     return null;
   }
-
-  const HAS_GRAPH = graphinRef.current?.graph && !graphinRef.current.graph.destroyed;
 
   const { renderComponents, InitializerComponent, InitializerProps, GICC_LAYOUT_COMPONENT, GICC_LAYOUT_PROPS } =
     getComponents({ ...state, HAS_GRAPH }, config.components, ComponentAssets);