+++ packages/gi-assets-basic/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @antv/gi-assets-basic
 
+## 2.3.5
+
+### Patch Changes
+
+- update assets
+- Updated dependencies
+  - @antv/gi-sdk@2.3.4
+
+## 2.3.4
+
+### Patch Changes
+
+- update assets
+
+## 2.3.3
+
+### Patch Changes
+
+- update assets
+- Updated dependencies
+  - @antv/gi-common-components@1.2.3
+  - @antv/gi-sdk@2.3.3
+
 ## 2.3.2
 
 ### Patch Changes