+++ packages/gi-assets-graphscope/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/gi-assets-graphscope
 
+## 2.0.8
+
+### Patch Changes
+
+- update assets
+- Updated dependencies
+  - @antv/gi-sdk@2.3.3
+
 ## 2.0.7
 
 ### Patch Changes