+++ packages/gi-assets-neo4j/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/gi-assets-neo4j
 
+## 2.0.4
+
+### Patch Changes
+
+- update assets
+- Updated dependencies
+  - @antv/gi-sdk@2.3.3
+
 ## 2.0.3
 
 ### Patch Changes