+++ packages/gi-assets-scene/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @antv/gi-assets-scene
 
+## 2.1.4
+
+### Patch Changes
+
+- update assets
+- Updated dependencies
+  - @antv/gi-sdk@2.3.4
+
+## 2.1.3
+
+### Patch Changes
+
+- update assets
+
+## 2.1.2
+
+### Patch Changes
+
+- update assets
+- Updated dependencies
+  - @antv/gi-sdk@2.3.3
+
 ## 2.1.1
 
 ### Patch Changes