+++ packages/gi-httpservice/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @antv/gi-httpservice
 
+## 1.1.2
+
+### Patch Changes
+
+- update assets
+
 ## 1.1.1
 
 ### Patch Changes